python-scraper/requirements.txt
linkedin-scraper>=2.11.2
selenium>=4.15.0
python-dotenv>=1.0.0
